# doc/api.yaml
openapi: 3.0.0
info:
  description: |-
    This concerns the "Web and Software Architecture" course to develop from scratch a web application that allows users to share photos of special moments.
    It allows users to like, comment on posts, ban other users, and so forth.

    Note: "Banning" and "setting a profile to private" are different. Setting a profile to public (private = false), means that any user except those whom he bans can see the user's profile, comment, like, etc without following the user. This is a feature much like "explore" in many social media. If a user wants only his followers to see his feeds except those banned, he should set it to private. Otherwise, even those who did not follow him and were not banned by him can see his feeds.
  title: Photo Sharing App
  version: 1.0.0
  contact:
    name: Yusupha Juwara
    url: https://github.com/YusuphaJuwara
    email: yusuphajuwara@gmail.com
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  parameters:
...
    userID:  #Arbitrary name for the definition that will be used to refer to it. Not necessarily the same as the parameter name.
      in: path
      name: user-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The user ID
    postID:
      in: path
      name: post-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The post ID
...
    banID:
      in: path
      name: ban-user
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: A banned user ID
    commentID:
      in: path
      name: comment-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: A comment ID
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: uuid    # optional, arbitrary value for documentation purposes
  responses:
    NoContent204:
      description: Successful.
    BadRequest400:
      description: |-
        The request was not compliant with the documentation (eg. missing fields, etc).
        Further info in server logs.
    UnAuth401:
      description: |-
        Access token is missing or invalid.
        The request was not authorized. The user must be authorized first to perform this request.
    Forbidden403:
      description: |-
        Trying to access a resource that the user should not. This is forbidden action.
        E.g., 
        a banned user trying to get an info about the user who banned him, 
        or trying to delete a comment when the user is not the one who placed the comment or on whose post the comment was placed, etc.
        Further info in server logs.
    NotFound404:
      description: Not found.
    InternalServerError500:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  schemas:
    ID:
      description: The unique identifier. It is 36 characters containing 4 "-". 
      type: string
      format: uuid

      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 123e4567-e89b-12d3-a456-426655440000
...
    date-time:
      description: The date and time the post/comment was posted.
      type: string
      format: date-time
      example: '2017-07-21T17:32:28Z'

      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      minLength: 20   
      maxLength: 20
...
    message:
      description: The message as a comment or caption on a post.
      type: string
      pattern: '(?=.*?[a-zA-Z])[^]'  # Any char including new lines of at least one alphabet.
      minLength: 2
      maxLength: 1000
      example: "Wow, that's a beautifull family photo!!!"
...
    ids-array:
      description: array of uuid's.
      type: array
      minItems: 0
      maxItems: 1000
      items:  
        $ref: '#/components/schemas/ID'
...
    comment: 
      description: details about comment.
      type: object
      properties:
        comment-id:   
          $ref: '#/components/schemas/ID'
        post-id:  
          $ref: '#/components/schemas/ID'
        user-id:  
          $ref: '#/components/schemas/ID'
        date-time:  
          $ref: '#/components/schemas/date-time'
        message:  
          $ref: '#/components/schemas/message'
...
security:
  - bearerToken: []
...
paths:
...
  /users/{user-id}/ban:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [Ban]
      operationId: getBanUsers
      summary: Get the user IDs of the users who were banned by the given user.
      description: |-
        Returns an array of user IDs of the users who were banned by the given user.
      responses:
        "200":
          description: An array of user IDs.
          content:
            application/json:
              schema:   
                $ref: '#/components/schemas/ids-array'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/ban/{ban-user}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/banID"
    put:
      tags: [Ban]
      operationId: banUser
      summary: A user bans another user.
      description: |-
        A user with the \"user-id\" bans another user with \"ban-user\", provided that he didn't already.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  $ref: "#/components/responses/Forbidden403" #no forbidden error.
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    delete:
      tags: [Ban]
      operationId: unbanUser
      summary: a user unbans another user.
      description: |-
        A user with the \"user-id\" unbans another user with \"ban-user\".
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  $ref: "#/components/responses/Forbidden403" #no forbidden error.
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
...
  /posts/{post-id}/comments/:
    parameters:
      - $ref: "#/components/parameters/postID"
    get:
      tags: [Comments]
      operationId: getPhotoComments   #getPostComments
      summary: Get all the comments of a given post.
      description: |-
        This returns the comments' metadata and the comments themselves. This includes the comments IDs, the users' IDs who made the comments, the post ID on which the comments were made, the date and time the comments were made, and the comment messages themselves.
      responses:
        '200':
          description: An array of comments of a Requested post.
          content:
            application/json:
              schema: 
                description: An array of comments.
                type: array
                minItems: 0
                maxItems: 1000
                items:  
                  $ref: '#/components/schemas/comment'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    post:   # prof panizzi
      tags: [Comments]
      operationId: commentPhoto # commentPost
      summary: Places a new comment and returns the newly created comment ID.
      description: |-
        Posts a new \"comment\" on the given post ID by the given user.
        The server will create a new unique ID, the client can find it in the response.
      # parameters:     # no need because the user ID can be inferred from the token.
      #   - name: user-id
      #     in: query
      #     required: true
      #     schema:
      #       $ref: "#/components/schemas/ID"
      #     description: Th user ID that must be used in the query string.
      requestBody:
        required: true
        content:
          application/json:
            schema:   
              $ref: '#/components/schemas/message'
      responses:
        "201":
          description: Comment created successfully
          content:
            application/json:
              schema:   
                $ref: '#/components/schemas/ID'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /comments/{comment-id}:
    parameters:
      - $ref: "#/components/parameters/commentID"
    get:
      tags: [Comments]
      operationId: getComment
      summary: Get a given comment.
      description: |-
        This returns the comment's metadata and the comment itself. This includes the comment ID, the user ID who made the comment, the post ID on which the comment was made, the date and time the comment was made, and the comment message itself.
      responses:
        '200':
          description: The requested comment.
          content:
            application/json:
              schema:   
                $ref: '#/components/schemas/comment'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    delete:
      tags: [Comments]
      operationId: uncommentPhoto   #uncommentPost
      summary: Deletes a given comment by the user inferred from the token.
      description: |-
        Deletes a given comment by the given user. The user is either the one who placed the comment or on whose post the comment was placed.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"